+++ src/app/Providers.tsx
@@ -1,8 +1,9 @@
 "use client";
 
 import { ReactNode } from "react";
-import { AuthProvider } from "@/hooks/useAuth";
+
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { AuthProvider } from "@/lib/auth";
 
 const queryClient = new QueryClient();
