+++ src/utils/imageUpload.ts
@@ -1,33 +1,34 @@
-import { storage, db } from '@/lib/firebase/config';
-import { ref, uploadBytes, getDownloadURL } from 'firebase/storage';
-import { doc, updateDoc } from "firebase/firestore";
+import { storage, auth } from "@/lib/firebase/config";
+import { ref, uploadBytes, getDownloadURL } from "firebase/storage";
+import { updateProfile } from "firebase/auth";
 
 export const handleImageUpload = async (
-  file: File,
-  userId: string,
-  showSnackbar: (message: string, type: 'success' | 'error' | 'info') => void
+  file: File,  showSnackbar: (message: string, type: "success" | "error" | "info") => void
 ) => {
-    try {
-    // Create a reference to the file location
-    const profileImageRef = ref(storage, `users/${userId}/profile-image`);
-    
-    // Upload the file
+  // Get current user
+  const currentUser = auth.currentUser;
+  
+  if (!currentUser) {
+    showSnackbar('You must be logged in to upload an image', 'error');
+    throw new Error('User not authenticated');
+  }
+
+  try {
+    const profileImageRef = ref(storage, `users/${currentUser.uid}/profile-image`);
     await uploadBytes(profileImageRef, file);
-    
-    // Get the download URL
     const downloadURL = await getDownloadURL(profileImageRef);
     
-    // Update the user's profile in Firestore
-    const userRef = doc(db, "users", userId);
-    await updateDoc(userRef, {
-      photoURL: downloadURL,
-      updatedAt: new Date().toISOString()
+    // Update user's photoURL if needed
+    await updateProfile(currentUser, {
+      photoURL: downloadURL
     });
 
     showSnackbar('Profile image updated successfully', 'success');
     return downloadURL;
+  
   } catch (error) {
-    showSnackbar('Failed to upload image', 'error');
+    console.error("Error uploading image:", error);
+    showSnackbar("Failed to upload image", "error");
     throw error;
   }
 };
